requirements.txt
aiogram==3.27.0
supabase
flask
python-dotenv
